fix: migration off agent product add: delete agent batch product

+++ app/Http/Controllers/AgentProductController.php
@@ -52,6 +52,31 @@ class AgentProductController extends Controller
             return response()->json($th, 201);
         }
     }
+    public function deleteBatchOfUserAgentProduct(Request $request)
+    {
+        try {
+            $data = json_decode($request, true);
+
+            $reqUserID = $request['UserID'];
+            $userCntrl = new UserController();
+
+            $userID = $userCntrl->getUserIdByTelegramID($reqUserID);
+            if ($userID == null) {
+                return response()->json(false, 201);
+            }
+            $selectedProductList = json_decode($request['selectedProductList'], true);
+            foreach ($selectedProductList as $value) {
+                $aa = json_decode($value, true);
+                $value = (array) $aa;
+                if($value['productCategoriesId'] != null) $this->deleteAgentProductByPrCatIDAndUserID($userID, $value['productCategoriesId']);
+            }
+            return response()->json(true, 200);
+        } catch (\Throwable $th) {
+            \Log::info("$th");
+
+            return response()->json($th, 201);
+        }
+    }
     public function createANewAgentProduct(Request $request)
     {
         try {
@@ -94,7 +119,22 @@ class AgentProductController extends Controller
     public function deleteAgentProduct($id)
     {
         try {
-            $agentProduct = AgentProduct::find($request->id);
+            $agentProduct = AgentProduct::find($id);
+            $agentProduct->delete();
+            return response()->json(true, 200);
+        } catch (\Throwable $th) {
+            \Log::info("throw $th");
+            return response()->json(false, 500);
+        }
+    }
+    public function deleteAgentProductByPrCatIDAndUserID($userID,$productCatId)
+    {
+        \Log::info("message $userID $productCatId");
+        try {
+            $agentProduct = AgentProduct::where('user_id', $userID)->where('product_categories_id', $productCatId)->first();
+            if(!$agentProduct){
+                return ;
+            }
             $agentProduct->delete();
             return response()->json(true, 200);
         } catch (\Throwable $th) {
